[wasmfx] Represent effect handlers in Turboshaft
In preparation of implementing wasm effect handlers, this extends
Turboshaft's CheckExceptionOp mechanism to represent the control-flow of
a resume instruction with one handler. This will be extended later to
support multiple handlers.

The resume instruction ends with a builtin call that performs the stack
switch. The call should be followed by a CheckExceptionOp, which ends
the basic block and creates a split in the control-flow: one path for
the regular return, and one for the handler block.

If a resume instruction also happens to be in a local try block, this
creates a three-way split in the control-flow: one path for the normal
return, one for the exception handler, and one for the effect handler.

Although this is very similar to exception handling, there are some
noteworthy differences:
- We keep track of the tag associated with a handler block, so that we
  can later emit it as part of the handler table. This is needed for the
  handler lookup of the suspend instruction.
- Effect handlers do not nest like exception handlers: they only apply
  to the resume instruction that has them as immediates. So we do not
  need to extend the CatchScope mechanism, we simply set a value in the
  assembler to signal that the next call needs to handle effects, and
  reset it afterwards.

Bug: 388533754
Change-Id: I3d869933da705c6ef41b7350860c58999758e87c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7003560
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103084}