[interpreter] Merge hole elision info on continue
for loops that are structured

  for (start; cond; next) { body }

currently have a single scope for

  body
  next

but that's wrong, since body isn't guaranteed to run to the end before
running next. There might be a `continue`. This considers `body` to be
"branchy" with any continue merging before next, as well as the body
end itself.

Bug: 450618029
Change-Id: I0156e1c02eeafad880bd324f1f5441023d53139d
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7030683
Auto-Submit: Toon Verwaest <verwaest@chromium.org>
Commit-Queue: Toon Verwaest <verwaest@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103064}