[snapshot] Handle multiple NativeContexts in context (de)serializer
Do not serialize the "next_context_link" field in NativeContext to
the snapshot as this field is only used for the weak list of all
NativeContexts in that isolate. We were already doing this but
only for the top/initial NativeContext. In the flakily failing test,
we were actually serializing/deserializing multiple NativeContexts,
because the test was using Realm.create() to create another
NativeContext which was stored in a global variable of that initial
NativeContext.

In the deserializer we then use AddToWeakNativeContextList() to add
the freshly created NativeContexts to the isolate's weak list of
native contexts in PostProcessNewObject(). Previously this only
happened for the top/initial NativeContext as well.

Bug: 450135650
Change-Id: Ifbf426e3e1afef373820b5aa4410bc67e1b0e31c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7027312
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103088}