[json] Detect some cases where reviver has only 2 args
This is good for a 36% reduction in run-time on the JSON.parse mode
where there is a reviver function.  Sadly it's still about
five times slower than the fast parser.

To make use of this, the user would have to switch to an
arrow function. This has two effects. Firstly it bans the
magic arguments object, so that we can safely skip passing
the third argument. Secondly, arrow functions don't have a
'this' object, which limits the reviver's ability to foil
optimizations by manipulating the deserialized object graph
during JSON.parse.

The relevant change would be:
- var deserialized = JSON.parse(str, reviver);
+ var deserialized = JSON.parse(str, (key, val) => reviver(key, val));

Change-Id: Iace72ecb12205fff3cbdb6eb742f49d0d5270d1a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7027411
Reviewed-by: Patrick Thier <pthier@chromium.org>
Commit-Queue: Erik Corry <erikcorry@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103087}