[wasmfx] Consolidate effect handlers in the Turboshaft interface
Extend the decoder and the wasm Turboshaft graph interface to support
the full semantics of an effect handler, in particular pushing the
captured continuation on the stack and branching to the target label.

In the decoder, this works in a similar way to exception handlers:
after decoding the resume instruction, the decoder pushes the tag
params and the automatic continuation param on the stack, and calls
the ResumeHandler interface function to handle the branch.

In the Turboshaft graph interface, CallAndMaybeCatchException is
extended to support effects. A call may need to catch effects,
exceptions, or both. In each case, this is represented by a
CheckExceptionOp after the call, with two or three successors.

Bug: 388533754
Change-Id: If8eaf76ae13cf838c752f709fb54c9695f064409
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7011734
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103092}