[wasm][metrics] Use designated initializers
Instead of having a constructor that takes exactly the struct fields,
use designated initializers.

Note that even though this is in the public API of V8, those structs
should not be created outside of V8, so we remove the constructor
without deprecating it first.

R=jkummerow@chromium.org, leszeks@chromium.org

Change-Id: Icbde69af54d0b8283ead53713aa0b5863b435f19
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7021911
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103205}