[tiering] Uniformly treat deopts which don't invalidate code
There are currently two situations where deopts don't invalidate code:
* deopt in osr code and the deopt is outside the loop
* deopt to be able to osr from maglev to turbofan

We should ensure that the deoptimizer and NotifyDeoptimized (a runtime
function which runs after the deoptimizer is done) agree on when code
needs to be discarded and when not.

Also, if code is not discarded, or if deopts don't affect the main code,
we shouldn't change the tiering state.

Drive-By:
* Read DeoptInfo only once and cache it.
* Reset tiering state only once.

Change-Id: If789b6bbcab55eccf76010585152b1593b72134e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7036595
Reviewed-by: Toon Verwaest <verwaest@chromium.org>
Commit-Queue: Toon Verwaest <verwaest@chromium.org>
Auto-Submit: Olivier Flückiger <olivf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103114}