[wasmfx] Implement suspend instruction
Implement the suspend instruction in Turboshaft. Like other
stack-switching instructions, this goes through a builtin that saves the
current state in the active stack, and switches to the target stack.

The target stack is found via a C call that iterates the chain of
stacks and searches for an effect handler with a matching tag.

Bug: 388533754
Change-Id: I78b11582bca99101592c5c1eea33fa854477dc81
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7008594
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103130}