[wasm-gc] Bump kMaxCanonicalTypes to 2M
This uses one of four remaining unused bits in ValueType to increase
the maximum number of canonical types by a factor of 2.

Reasoning why this particular maximum makes sense: it is reasonable
to support repeated decoding of a module containing a very large
recgroup. Since we check that we could add the recgroup as new
if required *before* checking whether it matches an existing
canonicalized recgroup, this scenario in the limit requires the
internal maximum to be exactly twice as high as the standardized
per-module maximum.

Bonus improvement: print details when hitting kMaxCanonicalTypes
to aid debugging.

Bug: 42203975
Change-Id: I9fea25b141f04bd20ca3cf5c58d39799ebc82221
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7047048
Auto-Submit: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Jakob Kummerow <jkummerow@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103170}