[heap] Fix bad DCHECK in scavenger visitor
The DCHECK with the type checks is safe to use because there are no
pinned objects when weak objects are treated weakly. However the
location of the DCHECK, before the branch to check if weakness is
enabled, meant it was applied even when pinned objects were present.
Moving the DCHECK into the branch fixes this.

Bug: 452650281
Change-Id: I96533670cd79cc730a85ef89c53d7d56617eda2d
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7049480
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Auto-Submit: Omer Katz <omerkatz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103177}