[json] Optimize BuildJsonArray
Avoid write barrier on the young elements backing store. Even though the
elements backing may be allocated in the young generation,
GetWriteBarrierMode() currently forces a write barrier when marking is
on.

Also avoid the initial stores of fillers as we already know which Tagged
to write.

Bug: 443678572
Change-Id: Iffff4570484ec5f8e553d3ad06f49e558e6fbf01
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7053405
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Patrick Thier <pthier@chromium.org>
Reviewed-by: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103183}