[execution, heap] Remove Isolate::InFastCCall()
Isolate::InFastCCall() simply returns whether the latest transition
from JS to native code was a regular CEntry or a fast API call. This
also used to be the only JS to native transition that could be a fast
API call. So Isolate::InFastCCall() would correctly report whether
there was a fast API call on the stack or not and this method was used
in this way. However, now that we allow JS calls from fast api calls,
it could return false even though there is an earlier fast API call
on the stack.

This CL removes this method to clean up this confusion. It also drops
--allow_allocation_in_fast_api_call as this is now fully supported
and enabled in V8.

Bug: 424905890
Change-Id: Ib1626b30e1e8436513648c906ed4476f3cb6f77c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6678594
Reviewed-by: Andreas Haas <ahaas@google.com>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103212}