[heap] Fix ScavengerWeakObjectRetainer
ScavengerWeakObjectRetainer assumed JSFinalizationRegistry objects are
only retained weakly in the dirty finalization registry list. Therefore
it assumed objects are always either old or in from space. However,
old-to-new references are strongified through the remembered set. As a
result we could also observe an object that is already in to space.
Whenever an object in to space was encountered, we assume it's in from
space and has a forwarding address. Trying to resolve the false
forwarding address results in an invalid object and crashes.

Bug: 453990034
Change-Id: I7e7692e0d6328ba38d1a0135a375b56ff19ca9a9
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7072732
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103291}