[wasm] Fix DCHECK failure when inlining br_table into JS
This is analogous to https://crrev.com/c/6618664, where we forgot to fix
it not just for try_table but also br_table.

The issue is that we may not call `length()` after the decode interface
call, since the latter might have left the decoder in a failed state
(e.g., due to bailing out during Wasm-in-JS-inlining), which trips a
DHCECK in the former (see linked CL, same thing there).

Fixed: 451240002
Change-Id: If353a334942b7d45e9f7c3e924b01ad7364b9c9b
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7072273
Auto-Submit: Daniel Lehmann <dlehmann@chromium.org>
Reviewed-by: Thibaud Michaud <thibaudm@chromium.org>
Commit-Queue: Daniel Lehmann <dlehmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103287}