Introduce base::FatalNoSecurityImpact
This is similar to FatalOOM and a variant of FATAL that makes it clear
that the failure does not have any security impact. This can be used to
filter out or ignore irrelevant crashes in automated vulnerability
discovery systems.

As a first step, FatalNoSecurityImpact is now used for the various
"invalid size" errors when attempting to construct objects that are
larger than their maximum allowed size.

Bug: 453649202
Change-Id: Ieb1b09a06fc1a617faf69e97219f59512f91a4c3
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7079450
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103320}