[regexp] Fix duplicate named capture groups
The previous assumption that non-participating capture groups is always
a single, strictly increasing interval was incorrect, as the
non-participating interval can have holes.
E.g. in /(?<a>.)|(?<b>.)(?:(?<c>.)|(?<d>.))/, when creating the capture
group for 'd', 'a' (index 0) and 'c' (index 2) are non-participating,
while 'b' (index 1) is participating. It is therefore a syntax error if
'd' is equal 'b'.
To fix this, non-participating capture groups are now a list of
intervals instead of a single interval.

Fixed: 450694353
Change-Id: If66857ba97ce8cac151ed0a66b4ac0185997a5c4
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7079108
Commit-Queue: Patrick Thier <pthier@chromium.org>
Reviewed-by: Jakob Linke <jgruber@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103329}