[wasm-custom-desc] Fix ref.null exactness
The type of a ref.null is not automatically exact if the immediate is an
indexed type. Only if the immediate specifies exactness is the result
type exact. Fixing this lets us enable most of the disabled custom
descriptors spec tests (after also avoiding an unrelated DCHECK failure
in ref.get_desc decoding).

R=jkummerow@chromium.org, mliedtke@chromium.org

Bug: 454401192, 403372470
Change-Id: Ib380b91b38ec5d360834fbf80caf06e6046dd2bb
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7076310
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103334}