[api, heap] Report HeapStats on OOM through crash keys
This CL reports the fields of HeapStats as crash keys on OOM errors.
For OOM errors those fields should now show up as
"v8-oom-<field-name>".

Managing crash keys has to be done by the embedder. This CL exposes
a new API method SetCrashKeyStringCallbacks which the embedder can
use to provide methods for allocation crash keys and setting values
on them. This API is a bit more flexible than the old one because
we had to create separate CrashKeyIds for each key. The new API
just takes a string literal as argument.

Bug: 389951203
Change-Id: I30c15351d284b1691ad2cf6e1da94d9e6c38a892
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7072272
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103364}