[maglev] Fix lazy deopt frame translation for return registers
When building the deopt frame for a lazy deopt, Maglev was incorrectly
skipping values located in return registers. We thought that these
values are not necessary, since they'll be clobbered by the return
values of the builtin, but the return values of the builtin are dropped
if there is an exception at the same time. We'll still skip the values
when they are not actually necessary via liveness modification.

This could cause crashes, for example in a for-of loop where the
iterator's next() method both throws an exception and triggers a
lazy deopt.

Fixed: 455191443
Change-Id: Id237c003b83d7bd77106e9cc5298e9705458f849
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7087027
Reviewed-by: Victor Gomes <victorgomes@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Auto-Submit: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103369}