[compiler][arm64] Fix Dup and Shuffle recognition
The 'dup and shuffle' optimization worked on the assumption that, for
swizzles, we could drop the second argument and replace it with the
result of the dup. This would result in the dup producing the eight
most significant bytes and the canonical shuffle would produce the
other eight.

However, because we are dealing with swizzles, it may be that the
instruction used to implement the canonical shuffle doesn't read from
two registers. In this case the dup is never read and so the top bytes
are not what we expect.

Bug: 455592080
Change-Id: Ic205b01dec1a82631b1de1fce2d98e82d9af3804
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7095236
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Reviewed-by: Marja Hölttä <marja@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103394}