Reland "Reland "Reland "heap: Release paged pools every K seconds"""
This fixes the issue with the --single-threaded mode. When
--single-threaded is on, the task needs access to the isolate to be able
to repost itself to the isolate's task runner. The fix attaches the task
to the CancellableTaskManager of the current isolate (not the
IsolateGroup), to make sure that the task never executes when the
isolate is gone.

Bug: 429538831
Change-Id: I044cff44c7d8b0fcf13ec0565c020f0415f4ffcd
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7095296
Auto-Submit: Anton Bikineev <bikineev@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103389}