[maglev] Avoid emitting veneer pool right after CallSelf
To record lazy deopt location (safepoints) in Maglev (and Turbofan),
right after a Call, we record a safepoint at the current PC
offset. Something like:

  __ Call(...)
  masm.RecordLazyDeoptPoint()

For this to be correct, it's important that veneer pools are not
emitted right after the Call, because the RecordLazyDeoptPoint would
then see the address of the last branch in the veneer pool rather than
the address of the Call.

This is solved usually by having a BlockPoolsScope in
MacroAssembler::Call, which basically emits the veneer pool before the
Call rather than after (if needed).

This was however missing for CallSelf in Maglev, since this uses Bl
directly rather than Call.

Fixed: 455135359
Change-Id: I52f0f946cb67405d8fd3b9cb5663d41d4e3917a1
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7096818
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Auto-Submit: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103413}