[json] Parser: Speedup next token checks
In many places in the JSON parser we know statically which token is most
likely to come next. For these cases we can have a fast check if the
expected token matches the next character in the input stream (under the
assumption that usual payloads don't contain whitespaces).
To do so we create a mapping from the enum values to their corresponding character value.

Bug: 455800270
Change-Id: Iab5cda695e0fa1d36b30ce4545f35fc5a1ae8593
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7096438
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Patrick Thier <pthier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103409}