[maglev] Avoid conversions in NodeInfos to have a single canonical
We can currently have conversion nodes as keys into the NodeInfos,
which means that we might not find alternatives because we're querying
them with a conversion instead of the canonical node (which would lead
to duplicating conversion nodes). This CL fixes that by making sure
that the key into the NodeInfos is always the canonical node.

Drive-by: add some checks that Alternatives have the expected type
(and fix the PROCESS_*_CONV bugs that this uncovered).

Drive-by: refactor FindInfo/TryGetInfo in KNA.

Bug: 431933185
Change-Id: I93cdffe905c1e0a6403483106fddcad85f4bdbd3
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7105921
Reviewed-by: Victor Gomes <victorgomes@chromium.org>
Auto-Submit: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Victor Gomes <victorgomes@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103453}