Use CHECK_NO_SECURITY_IMPACT for checking heap-size related flags
We currently crash when an invalid combination of heap-size flags is
specified. Instead of crashing with a regular CHECK, we now use a
CHECK_NO_SECURITY_IMPACT to make it clear that there are no security
consequences from this (as it's just a misconfiguration error).

Change-Id: I106dbfdf9fd5b50e9c67075439dfa61d71aeb40a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7112419
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103467}