[maglev] Let HoleyFloat64ToTagged canonicalize smis by default
.. since current use patterns around BuildCheckSmi rely on this.
The expectation is that if `BuildCheckSmi(value)` passes at runtime,
then the value is guaranteed to be a smi when accessed later (e.g.
when it is stored into a slot).

One example: Array ctor speculation that picks the PACKED_SMI_ELEMENTS
elements kind, and guards values with `BuildCheckSmi`.

Some workarounds for this can now be removed; I've added comments
there and we can attempt that separately.

Bug: 454485895
Change-Id: I7ef81083a50151ea2ddd71626f2d5af7bafaca5c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7100539
Commit-Queue: Jakob Linke <jgruber@chromium.org>
Reviewed-by: Victor Gomes <victorgomes@chromium.org>
Auto-Submit: Jakob Linke <jgruber@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103479}