[sbx] check code kind in deopt/interpreter data setters
The DeoptimizationDataOrInterpreterData field has lead to a type
confusion bug in the past when swapping an optimized code object with a
baseline code object.
While we're checking for this swap now, we also need to ensure that the
type of this fields fits with the code kind. If you can trigger one of
these setters on an object of the wrong kind, you'd be able to bypass
the existing checks again. Fortunately, this is only possible right now
with tracing enabled.

Fixed: 457641143
Change-Id: I35eb44a35b83771417234628ccac66d30de23890
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7116599
Commit-Queue: Stephen Röttger <sroettger@google.com>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103484}