Sparkplug+: support patchable baseline code
A bytecode handler for an operation is often a generic handler that
deal with different types of inputs. This make the baseline compiler
to generate a call to a generic builtin for that operation. The
generic builtin can deal with all types of inputs, but this makes the
builtin very large and introduces a lot of branches to distinguish
inputs' types, which hurts the performance.

Sparkplug+ introduces some small and specific handlers to deal with
one single type of input. For example, this CL introduces some small
monomorphic handlers for builtin LoadIC to deal with one specific
kind of IC. We will patch the baseline code to use these specific
handlers on IC misses dynamically.

Bug: chromium:429351411
Change-Id: I85f842ae4050eebbdd8d25d5e113deae743c95d0
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6627216
Commit-Queue: Xu, Hao A <hao.a.xu@intel.com>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103495}