[riscv] Use tail-undisturbed and avoid overwriting inputs
The `vcompress` function is allowed to overwrite tail elements if the
tail is set to agnostic. However, we wrote the code under the assumption
that the adjacent vector of the vector group wouldn't be touched. It's
likely that vector units wouldn't actually change that register, but we
can't rely on that.

In the same function we also overwrote input registers. This is now
fixed.

Change-Id: I22c9079cb13a76f11a92a75a3afc57f27c7ad56a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6989490
Commit-Queue: Florian Loitsch <floitsch@rivosinc.com>
Reviewed-by: Kasper Lund <kasperl@rivosinc.com>
Cr-Commit-Position: refs/heads/main@{#103529}