[JSPI][Drumbrake] Disabling JSPI for Wasm Interpreter and Sync
This PR disables the JSPI integration when running in wasm interpreter
(jitless mode). JSPI recently removed the flags required to be enabled,
which broke the previous disabling of this feature for drumbrake.

In this CL we also disable customs descriptors for Wasm Drumbrake.

Include in this CL build fixes in Drumbrake fuzzers from filename
change.

Bug: None
Change-Id: I975a8387af03245d47df11b1f94591eabc8fae0b
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7121798
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Commit-Queue: Emmanuel Romero Ruiz <emromero@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#103545}