[api] add Module::GetResourceName()
This allows embedders to retrieve the resource name of a module
directly instead of having to keep track of them on the side.
If the module is a Source Text Module, returns the name
that was passed by the embedder as resource_name to the ScriptOrigin.
If it's a Synthetic Module, returns the module_name passed to
CreateSyntheticModule().

Change-Id: I320c55a5c1eac0f2c5f14a1a8aa4773f3bc8a12c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7119198
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Joyee Cheung <joyee@igalia.com>
Cr-Commit-Position: refs/heads/main@{#103546}