Fix maglev regression test to not use seeded values
This test, introduced in crrev.com/c/5721508, was generated by a fuzzer.
It uses seeded randomness for identifying specific builtins and
properties, unfortunately this randomness breaks down every time a new
builtin is added.

There's no reason for it to use randomness. I updated it to instead use
the precise values returned by that seeded randomness in a non-Temporal
context.

This should fix the breakage on lkgr.

Change-Id: I172462406147d237f4d22bdd417cd5846a6a6964
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7126080
Reviewed-by: Rezvan Mahdavi Hezaveh <rezvan@chromium.org>
Commit-Queue: Manish Goregaokar <manishearth@google.com>
Cr-Commit-Position: refs/heads/main@{#103549}