[heap] Check whether gc-tracing category is enabled for tracing
For perfetto tracing we required a mix of additional --trace-gc
command line flags for emitting additional data about GC events.
In this CL we remove this and just check whether the v8.gc
tracing category is enabled. This is a simple pointer dereference, so
it will be cheap to check.

Bug: 456384548
Change-Id: Idde69145dd5e96db4d31d679d49111ec9b7e7f0e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7123829
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103564}