[riscv] Fix loading of is_marking_flag
This patch is to fix the issue which reported in test/mjsunit/whitespaces.js.

The is_marking_flag field has type uint8_t, so it should be loaded using the
`lbu` instruction instead of `ld`. Using `ld` to load the data and comparing it
with zero may lead to incorrect branching in JumpIfMarking and JumpIfNotMarking.

Change-Id: I8c3b861fd67bc93e1bfff30352667488db2c6bdb
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7126739
Commit-Queue: Yahan Lu (LuYahan) <yahan@iscas.ac.cn>
Reviewed-by: Kasper Lund <kasperl@rivosinc.com>
Reviewed-by: Yahan Lu (LuYahan) <yahan@iscas.ac.cn>
Cr-Commit-Position: refs/heads/main@{#103585}