[wasm-custom-desc] Don't call .instance_descriptors() for Wasm maps
Wasm maps with Custom Descriptors use the "instance descriptors" slot
for storing their descriptor object, but its type is WasmStruct, not
DescriptorArray, so we must be careful not to mix them up.
This patch fixes the specific issue that the fuzzer found, and adds
a DCHECK that flushed out another instance of the pattern in heap
verification, and includes the results of a manual audit that found
two more instances in object printing and heap snapshotting.

Bug: 403372470
Change-Id: Ia50d2b941895f677058df77125cc0cc28d1f362a
Fixed: 457106696
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7131198
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Auto-Submit: Jakob Kummerow <jkummerow@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103645}