[trace] Don't add 'new' to Wasm call site
When constructors builtins are inlined in Maglev, but the
constructor itself is not, we need to mark the call site
flag kIsConstructor. This allows the stack trace builder to add
the 'new' string.

However, this should not be done if the target is Wasm.
This creates the same stack trace string as Ignition.

Fixed: 450308355
Change-Id: I5c21eba3b46f97ddca9c51c21a66aa8169c0293c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7136920
Reviewed-by: Marja Hölttä <marja@chromium.org>
Auto-Submit: Victor Gomes <victorgomes@chromium.org>
Commit-Queue: Victor Gomes <victorgomes@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103617}