[icu] Fix `LookupAndValidateUnicodeExtensions()` for ICU 77
ICU 77 `getKeywordValue()` changed to return
`U_STRING_NOT_TERMINATED_WARNING` when the value is exactly
the same size as the given buffer size. In this case, the
buffer is not null-terminated[1].

This patch fixes `LookupAndValidateUnicodeExtensions()` for
this change.

[1] https://unicode-org.github.io/icu/userguide/strings/#using-c-strings-nul-terminated-vs-length-parameters

Bug: 458186954
Change-Id: Iaff890f93cc392d4deb71ab881ab4bcce97f5b34
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7137918
Auto-Submit: Koji Ishii <kojii@chromium.org>
Commit-Queue: Jakob Linke <jgruber@chromium.org>
Reviewed-by: Jakob Linke <jgruber@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103662}