[maglev] Silence nan's when inlining the Array ctor
.. except when they should not be silenced. The latter is the case when
element values can be an "internal" hole nan value; in other words, when
we've generated the hole nan value internally as a representation of a
HOLEY_DOUBLE_ELEMENTS hole. This can currently happen for the 0-arg
Array ctor (when we preallocate elements); the 1-arg case, which results
in an empty holey array; and when creating an array from a boilerplate
object.

In all other cases, silence NaNs.

With this fix, we also re-enable the double Array ctor inlining.

Bug: 457866804
Change-Id: I4a2ecaa4c3eaa61e8d9df645ddf31fa05072bd91
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7137283
Reviewed-by: Victor Gomes <victorgomes@chromium.org>
Commit-Queue: Jakob Linke <jgruber@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103649}