[wasm-custom-desc] Exact function imports
We previously changed ref.func to produce an inexact type when
its input is an imported function, because importing allows subtypes.
That regresses modularity though, because it makes imported functions
less optimizable than functions defined in the module. The solution is
to introduce a new kind of import: exact function. Providing a value
whose actual type is a subtype of an exact import is disallowed.

Bug: 403372470
Change-Id: I1c631140acfcf09a6cbbd5dd37079470600c9435
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7124024
Reviewed-by: Daniel Lehmann <dlehmann@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103650}