[wasmfx] Visit active stack's func ref and cont
Wasm's StackMemory holds two references: the current cont ref associated
with the stack, and the func ref that is called when the stack is
resumed for the first time.
These objects are visited as roots in StackMemory::Iterate, but this
function was only called on inactive stacks. Also call it on the active
stack to ensure that we visit these special stack roots.

R=jkummerow@chromium.org

Fixed: 455832038
Change-Id: I9147058c06305ec42385a5781ec8604271fc4289
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7146381
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103699}