[x64] Use the flags set by -march for snapshot.
When generating the builtins for the snapshot we might as well
make use of any CPU features implied by the -march flag of the
C++ compiler. The resulting binary will only work on CPUs that
support these flags, so there's no harm in using them in the
builtins.

Change-Id: I23f816e6584c58102633c8c9fbdc7a62905a7f5a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7151303
Reviewed-by: Erik Corry <erikcorry@chromium.org>
Commit-Queue: Erik Corry <erikcorry@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103716}