[heap] Remove ArrayBufferList::bytes_
Turns out this field was complex to track but we didn't use it
for anything. So this CL removes this field entirely. For tests we
can simply iterate the ArrayBufferExtensions and sum up their
lengths.

Now that this field is gone, it cannot overflow anymore either like
in the linked bug on 32-bit.

Bug: 460166671
Change-Id: I15e50c60f0f612eb0bb0aab6c7cc5a99e501181a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7156719
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103736}