[turboshaft] Cleanup ConvertUntaggedToJSPrimitiveOrDeopt
It only accepts Word inputs and only produces Smis; so this CL renames
it to ConvertWordToSmiOrDeopt to that it's clearer what it does.

If we ever need it to support more things, we can always rename and
re-introduce the JSPrimitiveKind, but for now this will make it a bit
simpler/clearer.

Change-Id: I47e5d355ed0032d185626b17a4552ce78ff84704
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7131740
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Auto-Submit: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103743}